requirements.txt
discord.py==2.3.2
yt-dlp==2024.3.10
python-dotenv==1.0.0
spotipy==2.23.0
soundcloud-v2==1.3.3
deezer-python==5.10.0
asyncio==3.4.3
PyNaCl==1.5.0